+++ client/web/src/site/NeedsRepositoryConfigurationAlert.tsx
@@ -1,6 +1,7 @@
 import classNames from 'classnames'
-import * as React from 'react'
-import { Link } from 'react-router-dom'
+import React from 'react'
+
+import { Link } from '@sourcegraph/shared/src/components/Link'
 
 import { DismissibleAlert } from '../components/DismissibleAlert'
 import { eventLogger } from '../tracking/eventLogger'
@@ -13,9 +14,7 @@ const onClickCTA = (): void => {
  * A global alert telling the site admin that they need to configure repositories
  * on this site.
  */
-export const NeedsRepositoryConfigurationAlert: React.FunctionComponent<{ className?: string }> = ({
-    className = '',
-}) => (
+export const NeedsRepositoryConfigurationAlert: React.FunctionComponent<{ className?: string }> = ({ className }) => (
     <DismissibleAlert
         partialStorageKey="needsRepositoryConfiguration"
         className={classNames('alert-success d-flex align-items-center', className)}
