+++ cmd/frontend/graphqlbackend/site_alerts.go
@@ -12,7 +12,6 @@ import (
 	"github.com/inconshreveable/log15"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/globals"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/updatecheck"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
@@ -105,9 +104,6 @@ func init() {
 
 	// Notify when updates are available, if the instance can access the public internet.
 	AlertFuncs = append(AlertFuncs, updateAvailableAlert)
-
-	// Notify about postgres deprecation
-	AlertFuncs = append(AlertFuncs, deprecationAlert)
 
 	// Notify admins if critical alerts are firing, if Prometheus is configured.
 	prom, err := srcprometheus.NewClient(srcprometheus.PrometheusURL)
@@ -229,25 +225,6 @@ func outOfDateAlert(args AlertFuncArgs) []*Alert {
 		return nil
 	}
 	return []*Alert{alert}
-}
-
-// This should be removed from 3.27
-func deprecationAlert(args AlertFuncArgs) []*Alert {
-	if envvar.SourcegraphDotComMode() {
-		return nil
-	}
-
-	cv, err := semver.NewVersion(version.Version())
-	if err != nil {
-		log15.Error("cannot determine version", "error", err)
-		return nil
-	}
-
-	if cv.Minor() == 26 && args.IsSiteAdmin {
-		return []*Alert{{TypeValue: AlertTypeInfo, MessageValue: "Support for Postgres v11.x and below will be deprecated from Sourcegraph v3.27. Please reach out to support@sourcegraph.com if you require assistance upgrading.", IsDismissibleWithKeyValue: "true"}}
-	}
-
-	return nil
 }
 
 func determineOutOfDateAlert(isAdmin bool, months int, offline bool) *Alert {
